+++ lazyros/lazyros.py
@@ -13,19 +13,19 @@ from textual.widgets import (
     TabPane,
 )
 
-from lazyros.widgets.node.node_list_widget import NodeListWidget
-from lazyros.widgets.node.log_view_widget import LogViewWidget
-from lazyros.widgets.node.info_view_widget import InfoViewWidget
-from lazyros.widgets.node.lifecycle import LifecycleWidget
+from lazyros.widgets.node.node_list import NodeListWidget
+from lazyros.widgets.node.node_log import LogViewWidget
+from lazyros.widgets.node.node_info import InfoViewWidget
+from lazyros.widgets.node.node_lifecycle import LifecycleWidget
 
-from lazyros.widgets.topic.info_view_widget import TopicInfoWidget
-from lazyros.widgets.topic.topic_list_widget import TopicListWidget
-from lazyros.widgets.parameter.parameter_list_widget import ParameterListWidget
-from lazyros.widgets.topic.echo_view_widget import EchoViewWidget
-from lazyros.widgets.parameter.parameter_value_widget import ParameterValueWidget
-from lazyros.widgets.parameter.parameter_info_widget import ParameterInfoWidget
-from lazyros.modals.topic_info_modal import TopicInfoModal
-from lazyros.modals.message_modal import MessageModal
+from lazyros.widgets.topic.topic_info import TopicInfoWidget
+from lazyros.widgets.topic.topic_list import TopicListWidget
+from lazyros.widgets.topic.topic_echo import EchoViewWidget
+
+from lazyros.widgets.parameter.parameter_list import ParameterListWidget
+from lazyros.widgets.parameter.parameter_value import ParameterValueWidget
+from lazyros.widgets.parameter.parameter_info import ParameterInfoWidget
+
 from lazyros.utils.utility import ros_spin_thread, signal_shutdown
 
 from textual.screen import ModalScreen
