+++ lazyros/widgets/node/node_info.py
