+++ lazyros/widgets/node/node_log.py
